fix(gui): close the re-review findings on the Profiles fixes
The re-review found the fix batch had left one arm of the defect
it was fixing, and had traded two fail-opens for two stale reads.

**The fallback arm still asked half the rule.** `profileVerdict`'s
`.none` case answered `StandardProfiles.standard(for:)`, but the
live path answers `composeMonitorChangeFallback`, which recomposes
the beginner LADDER while on the starter baseline (#485) — and
`StarterLadder` is `isStandard: false`, so the simpler query can
never name it. The card said "the built-in Dual Developer" while
Starter composed, with the profile header saying "Starter" two
cards up. It now routes through the same fallback.

**The separate-Spaces snapshot baked in the display count.**
`recommendsSharedSpaces(displayCount:)` is `count > 1 &&
hasSeparateSpaces()`, and nothing refreshes the dashboard on a
display change — so plugging in a second monitor left the binding
rows live with the stale one-display answer, the very fail-open
the previous commit set out to close. Only the preference is
snapshotted now; the count is read live.

**The verdict sentence was assembled from two moments.** A live
`displays.count` beside a snapshotted verdict could read "2
screens → Desk (these exact monitors)" about a one-screen match.
Count and verdict now ride in one `ProfileResolution` value.

Also: `presetsApply` declared one arm of its two-arm predicate, so
it gains `presetApplyUnavailable` and the register grows to four —
`screenCountMismatch` and `editingStoredProfile` are deleted,
being tags no census key declares (pre-release, no dead slots).
Two docstrings corrected to what their guards actually hold, and
`SettingsModel` splits its backend sync out at the file ceiling.
