+++ Sources/KiwiDesk/Settings/Census/SettingKey+Profiles.swift
@@ -53,7 +53,7 @@ extension ProfilesKey {
                 .profiles,
                 .presets,
                 .atRest,
-                gate: .runtime(.screenCountMismatch)
+                gate: .runtime(.presetApplyUnavailable)
             )
         }
     }
